heredoc progression

+++ exe/ft_before_exe.c
@@ -6,7 +6,7 @@
 /*   By: mbriand <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/09 01:31:21 by mbriand           #+#    #+#             */
-/*   Updated: 2024/07/03 03:00:51 by mbriand          ###   ########.fr       */
+/*   Updated: 2024/07/03 18:10:37 by mbriand          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int	ft_if_exitcode_2(t_minishell *data, t_commands *p_cmd)
 
 int	ft_if_no_command(t_minishell *data, t_commands *p_cmd)
 {
-	if (!p_cmd->cmd)
+	if (!p_cmd->cmd && !p_cmd->next)
 	{
 		if (p_cmd->hd_stop)
 			unlink(p_cmd->heredoc);
